+++ CHANGELOG.md
@@ -12,7 +12,7 @@
   The new option `dir_browser.davmount_links` controls link display (default: false).
 - #222 Discrepancy between "getetag" property and ETag header
 - TODO: Support LibreOffice in dir_browser
-- TODO: #185 FileLikeQueue not compatible with Python3
+- TODO: #185 FileLikeQueue USE BYTES
 
 ## 3.1.2 / Unreleased